# train/swagger.yaml
openapi: 3.0.0
info:
  title: AI Asset Valuation API
  version: 1.0.0
paths:
  /api/analysis/upload-and-analyze:
    post:
      summary: Upload and analyze images
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
  /api/reports:
    post:
      summary: Create new property report
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyReportCreate'
      responses:
        '200':
          description: Successful response
    get:
      summary: List user's property reports
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Successful response
  /api/reports/{report_id}:
    get:
      summary: Get detailed report
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
components:
  schemas:
    PropertyReportCreate:
      type: object
      properties:
        address:
          type: string
        property_type:
          type: string
        land_area:
          type: number
        usable_area:
          type: number
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        floors:
          type: integer
        direction:
          type: string
        legal_status:
          type: string
        furniture:
          type: string
        width:
          type: number
        length:
          type: number
        overall_condition:
          type: string
        cleanliness:
          type: string
        maintenance_status:
          type: string
        major_issues:
          type: array
          items:
            type: string
        overall_description:
          type: string
        images:
          type: array
          items:
            type: object
            properties:
              filename:
                type: string
              url:
                type: string
              key:
                type: string
        ai_analysis_raw:
          type: object
    AnalysisResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        images:
          type: array
          items:
            type: object
            properties:
              filename:
                type: string
              url:
                type: string